+++ test/fixtures/basic-import/expected.es2015.js
@@ -2,6 +2,6 @@
 
 function _interopRequireWildcard(obj) {}
 
-var testModule = Promise.resolve().then(function () {
-  return _interopRequireWildcard(require('test-module'));
+var testModule = Promise.resolve('test-module').then(function (_) {
+  return _interopRequireWildcard(require(_));
 });