+++ test/fixtures/chained-import/expected.js
@@ -1,5 +1,5 @@
 function _interopRequireWildcard(obj) {}
 
-Promise.resolve().then(() => _interopRequireWildcard(require('test-module'))).then(() => Promise.resolve().then(() => _interopRequireWildcard(require('test-module-2'))));
+Promise.resolve('test-module').then(_ => _interopRequireWildcard(require(_))).then(() => Promise.resolve('test-module-2').then(_ => _interopRequireWildcard(require(_))));
 
-Promise.all([Promise.resolve().then(() => _interopRequireWildcard(require('test-1'))), Promise.resolve().then(() => _interopRequireWildcard(require('test-2'))), Promise.resolve().then(() => _interopRequireWildcard(require('test-3')))]).then(() => {});
+Promise.all([Promise.resolve('test-1').then(_ => _interopRequireWildcard(require(_))), Promise.resolve('test-2').then(_ => _interopRequireWildcard(require(_))), Promise.resolve('test-3').then(_ => _interopRequireWildcard(require(_)))]).then(() => {});