+++ test/fixtures/dynamic-argument/expected.js
@@ -2,5 +2,5 @@ function _interopRequireWildcard(obj) {}
 
 const MODULE = Object('test-module');
 
-Promise.resolve().then(() => _interopRequireWildcard(require(`${MODULE}`)));
-Promise.resolve().then(() => _interopRequireWildcard(require(`test-${MODULE}`)));
+Promise.resolve(`${MODULE}`).then(_ => _interopRequireWildcard(require(_)));
+Promise.resolve(`test-${MODULE}`).then(_ => _interopRequireWildcard(require(_)));