+++ test/fixtures/import-with-comment/expected.noInterop.js
@@ -1,2 +1,2 @@
-Promise.resolve().then(() => require('my-module'));
-Promise.resolve().then(() => require('my-module'));
+Promise.resolve('my-module').then(_ => require(_));
+Promise.resolve('my-module').then(_ => require(_));