+++ test/fixtures/nested-import/expected.js
@@ -1,7 +1,7 @@
 function _interopRequireWildcard(obj) {}
 
 function getModule(path) {
-  return Promise.resolve().then(() => _interopRequireWildcard(require('test-module')));
+  return Promise.resolve('test-module').then(_ => _interopRequireWildcard(require(_)));
 }
 
 getModule().then(() => {});