+++ test/fixtures/nested-import/expected.noInterop.js
@@ -1,5 +1,5 @@
 function getModule(path) {
-  return Promise.resolve().then(() => require('test-module'));
+  return Promise.resolve('test-module').then(_ => require(_));
 }
 
 getModule().then(() => {});