+++ test/fixtures/non-string-argument/expected.js
@@ -1,10 +1,10 @@
 function _interopRequireWildcard(obj) {}
 
-Promise.resolve().then(() => _interopRequireWildcard(require(`${{ 'answer': 42 }}`)));
-Promise.resolve().then(() => _interopRequireWildcard(require(`${['foo', 'bar']}`)));
-Promise.resolve().then(() => _interopRequireWildcard(require(`${42}`)));
-Promise.resolve().then(() => _interopRequireWildcard(require(`${void 0}`)));
-Promise.resolve().then(() => _interopRequireWildcard(require(`${undefined}`)));
-Promise.resolve().then(() => _interopRequireWildcard(require(`${null}`)));
-Promise.resolve().then(() => _interopRequireWildcard(require(`${true}`)));
-Promise.resolve().then(() => _interopRequireWildcard(require(`${Symbol()}`)));
+Promise.resolve(`${{ 'answer': 42 }}`).then(_ => _interopRequireWildcard(require(_)));
+Promise.resolve(`${['foo', 'bar']}`).then(_ => _interopRequireWildcard(require(_)));
+Promise.resolve(`${42}`).then(_ => _interopRequireWildcard(require(_)));
+Promise.resolve(`${void 0}`).then(_ => _interopRequireWildcard(require(_)));
+Promise.resolve(`${undefined}`).then(_ => _interopRequireWildcard(require(_)));
+Promise.resolve(`${null}`).then(_ => _interopRequireWildcard(require(_)));
+Promise.resolve(`${true}`).then(_ => _interopRequireWildcard(require(_)));
+Promise.resolve(`${Symbol()}`).then(_ => _interopRequireWildcard(require(_)));