+++ test/fixtures/non-string-argument/expected.noInterop.js
@@ -1,8 +1,8 @@
-Promise.resolve().then(() => require(`${{ 'answer': 42 }}`));
-Promise.resolve().then(() => require(`${['foo', 'bar']}`));
-Promise.resolve().then(() => require(`${42}`));
-Promise.resolve().then(() => require(`${void 0}`));
-Promise.resolve().then(() => require(`${undefined}`));
-Promise.resolve().then(() => require(`${null}`));
-Promise.resolve().then(() => require(`${true}`));
-Promise.resolve().then(() => require(`${Symbol()}`));
+Promise.resolve(`${{ 'answer': 42 }}`).then(_ => require(_));
+Promise.resolve(`${['foo', 'bar']}`).then(_ => require(_));
+Promise.resolve(`${42}`).then(_ => require(_));
+Promise.resolve(`${void 0}`).then(_ => require(_));
+Promise.resolve(`${undefined}`).then(_ => require(_));
+Promise.resolve(`${null}`).then(_ => require(_));
+Promise.resolve(`${true}`).then(_ => require(_));
+Promise.resolve(`${Symbol()}`).then(_ => require(_));